Changed DBColumn columns' value to be nullable (#101)
Prior to this, the column does not support null value, leading us have
to use NonNull column value, and in fact, we may need Nullable value in
DB and the puts method of ContentValues support Nullable value, Maybe
you'll say emptyColumn can do it, but actually it's hard to use it. So
I use Nullable instead of Nullable for the value argument. It will be
convenient.